Tweaks to upgrade manager
